docs: moved predicate transactions docs snippets
