+++ apps/docs/src/guide/predicates/custom-transactions.md
@@ -6,8 +6,8 @@ Custom transactions can be shaped via a `ScriptTransactionRequest` instance. For
 
 However, this guide will demonstrate how to use a predicate in a custom transaction. Consider the following predicate, where a configurable pin must be used to validate the predicate and unlock the funds:
 
-<<< @/../../docs-snippets/test/fixtures/forc-projects/configurable-pin/src/main.sw#predicate-with-configurable-pin-1{rust:line-numbers}
+<<< @/../../docs-snippets2/sway/configurable-pin/src/main.sw#full{rust:line-numbers}
 
 We can interact with the above and include it in a custom transaction like so:
 
-<<< @/../../docs-snippets/src/guide/predicates/predicate-custom-transactions.test.ts#predicate-custom-transaction{ts:line-numbers}
+<<< @/../../docs-snippets2/src/predicates/custom-transactions.ts#predicate-custom-transaction{ts:line-numbers}
